fix: replacing + symbols with - to generate valid aws s3 urls from object name
